[lgthinq][fix] fix code to remove warnings in complication

+++ bundles/org.openhab.binding.lgthinq/src/main/java/org/openhab/binding/lgthinq/lgservices/model/devices/ac/ACOpMode.java
@@ -37,7 +37,7 @@ public enum ACOpMode {
     }
 
     public static ACOpMode statusOf(int value) {
-        switch ((int) value) {
+        switch (value) {
             case 0:
                 return COOL;
             case 1:
